add some sanity checks to make sure modules are correctly added
